requirements/base.txt
django-authtools==1.6.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django==2.0.9
easy-thumbnails==2.5
pillow==5.3.0
pytz==2018.7
simple-salesforce==0.74.2
djangorestframework==3.9.4
django-rest-auth==0.6.0
requests==2.21.0
psycopg2>=2.7.7
social-auth-app-django==3.1.0
social-auth-core==3.0.0
Celery==4.3.0
django-celery-results==1.1.2